requirements.txt
tensorflow-macos>=2.13.0
tensorflow-metal>=1.0.0
numpy>=1.24.0
pillow>=10.0.0
